lib/chroma_visualise/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ally-ai-chroma-visualise"
version = "0.1.0"
authors = [{ name = "Tugbay Atilla", email = "tugbayatilla@gmail.com" }]
description = "visualise package for ally chroma"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = ["ally-ai-chroma>=0.3.0", "umap-learn", "matplotlib"]

[project.urls]
Homepage = "https://github.com/users/tugbayatilla/projects/3/"
Issues = "https://github.com/users/tugbayatilla/projects/3/views/8"
Repository = "https://github.com/tugbayatilla/ally_ai.git"
